rename project package to bittium
